+++ Gag.py
@@ -18,12 +18,11 @@ from math import floor as math_floor
 from typing import List, Optional
 
 from core.AttackGlobals import GROUP
-
-from Exceptions import (GagCountError, LockedGagError, LockedGagTrackError,
-                         NotEnoughGagsError)
+from core.Exceptions import (GagCountError, LockedGagError,
+                             LockedGagTrackError, NotEnoughGagsError)
 from GagGlobals import (DEFAULT_GAG_COUNT, DEFAULT_TRACK_EXPS_CURRENT, GAG,
-                         GAG_CARRY_LIMITS, GAG_DAMAGE, GAG_LABELS,
-                         GAG_TRACK_LABELS, LEVELS, MULTI_TARGET_GAGS, TRACK)
+                        GAG_CARRY_LIMITS, GAG_DAMAGE, GAG_LABELS,
+                        GAG_TRACK_LABELS, LEVELS, MULTI_TARGET_GAGS, TRACK)
 
 
 def get_default_gag_count() -> List[List[int]]:
